Added regression tests to all the synthetic data sets created from .Rmd files.

+++ R/Diff4SynDataSets.R
@@ -2,21 +2,25 @@
 #'
 #' @param dirname the root name of the directories to diff.
 #'
-#' @return The return value of diff command.
+#' @param unlink if \code{TRUE} unlink \code{tmpdirname}, but do not unlink
+#' if there are diffs.
+#'
+#' @return Invisibly, the return value of the diff command.
 #'
 #' @export
 
-Diff4SynDataSets <- function(dirname) {
+Diff4SynDataSets <- function(dirname, unlink) {
   regressdirname <- paste0("long.test.regression.data/", dirname)
   if (!file.exists(regressdirname)) stop(regressdirname, " does not exist")
   tmpdirname <- paste0("tmp.", dirname)
   if (!file.exists(tmpdirname)) stop(tmpdirname, " does not exist")
-  cmd <- paste("diff -rq", tmpdirname, regressdirname)
-  cmd.result <- system(cmd)
+  cmd.result <- system2("diff", c("-rq", tmpdirname, regressdirname))
   if (cmd.result == 0) {
     # No differences
-    unlink.res <- unlink(tmpdirname, recursive = TRUE, force = TRUE)
-    if (unlink.res != 0) warning("failed to unlink ", tmpdirname)
+    if (unlink) {
+      unlink.res <- unlink(tmpdirname, recursive = TRUE, force = TRUE)
+      if (unlink.res != 0) warning("failed to unlink ", tmpdirname)
+    }
   }
-  return(cmd.result)
+  invisible(cmd.result)
 }
